+++ src/texto/extensions/task-item/view/svg/itemIcon.svgnode.tsx
@@ -4,7 +4,8 @@ export const itemIconNode = litView.node(function* (params: {iconId?: string}) {
 	while (true) {
 		params = yield (
 			<svg>
-				<use href={`#${params.iconId}`} />
+				{/* href (SVG2) + xlink:href (legacy) for max <use> compatibility */}
+				<use href={`#${params.iconId}`} xlinkHref={`#${params.iconId}`} />
 			</svg>
 		);
 	}